+++ packages/kit/types.internal.d.ts
@@ -1,4 +1,13 @@
-import { Adapter, GetContext, GetSession, Handle, Incoming, Load, Response } from './types';
+import {
+	Adapter,
+	GetContext,
+	GetSession,
+	Handle,
+	Incoming,
+	Load,
+	RequestHandler,
+	Response
+} from './types';
 import { UserConfig as ViteConfig } from 'vite';
 import { Response as NodeFetchResponse } from 'node-fetch';
 
@@ -7,6 +16,8 @@ declare global {
 		env: Record<string, string>;
 	}
 }
+
+type PageId = string;
 
 export type Logger = {
 	(msg: string): void;
@@ -56,14 +67,16 @@ export type ValidatedConfig = {
 
 export type App = {
 	init: ({
-		paths
+		paths,
+		prerendering
 	}: {
 		paths: {
 			base: string;
 			assets: string;
 		};
+		prerendering: boolean;
 	}) => void;
-	render: (incoming: Incoming, options: SSRRenderOptions) => ResponseWithDependencies;
+	render: (incoming: Incoming, options: SSRRenderOptions) => Response;
 };
 
 // TODO we want to differentiate between request headers, which
@@ -71,10 +84,6 @@ export type App = {
 // 'set-cookie' is a `string[]` (or at least `string | string[]`)
 // but this can't happen until TypeScript 4.3
 export type Headers = Record<string, string>;
-
-export type ResponseWithDependencies = Response & {
-	dependencies?: Record<string, Response>;
-};
 
 export type Page = {
 	host: string;
@@ -134,17 +143,16 @@ export type SSRPage = {
 	type: 'page';
 	pattern: RegExp;
 	params: GetParams;
-	parts: SSRPagePart[];
-	style: string;
-	css: string[];
-	js: string[];
+	parts: PageId[];
 };
 
 export type SSREndpoint = {
 	type: 'endpoint';
 	pattern: RegExp;
 	params: GetParams;
-	load: () => Promise<any>; // TODO
+	load: () => Promise<{
+		[method: string]: RequestHandler;
+	}>;
 };
 
 export type SSRRoute = SSREndpoint | SSRPage;
@@ -157,8 +165,8 @@ export type CSRRoute = CSREndpoint | CSRPage;
 
 export type SSRManifest = {
 	assets: Asset[];
-	layout: SSRComponentLoader;
-	error: SSRComponentLoader;
+	layout: string;
+	error: string;
 	routes: SSRRoute[];
 };
 
@@ -177,18 +185,25 @@ export type SSRRenderOptions = {
 	local?: boolean;
 	template?: ({ head, body }: { head: string; body: string }) => string;
 	manifest?: SSRManifest;
+	load_component?: (
+		id: PageId
+	) => Promise<{
+		module: SSRComponent;
+		entry: string; // client-side module corresponding to this component
+		css: string[];
+		js: string[];
+		styles: string[];
+	}>;
 	target?: string;
 	entry?: string;
 	root?: SSRComponent['default'];
 	hooks?: Hooks;
 	dev?: boolean;
 	amp?: boolean;
+	dependencies?: Map<string, Response>;
 	only_render_prerenderable_pages?: boolean;
-	app_dir?: string;
-	get_component_path?: (id: string) => string;
 	get_stack?: (error: Error) => string;
 	get_static_file?: (file: string) => Buffer;
-	get_amp_css?: (dep: string) => string;
 	fetched?: string;
 	initiator?: SSRPage;
 	ssr?: boolean;